+++ src/Forms/Fields/Html.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace Thinktomorrow\Chief\Forms\Fields;
 
+use Thinktomorrow\Chief\Forms\Fields\Concerns\HasCharacterCount;
 use Thinktomorrow\Chief\Forms\Fields\Concerns\HasRedactorToolbar;
 
 class Html extends Component implements Field
 {
+    use HasCharacterCount;
     use HasRedactorToolbar;
 
     protected string $view = 'chief-form::fields.html';