+++ src/Forms/UI/views/fields/text.blade.php
@@ -11,10 +11,12 @@
         :autofocus="$hasAutofocus()"
         :attributes="$attributes
             ->merge($getCustomAttributes())
-            ->merge([$getWireModelType() => $getWireModelValue($locale ?? null)])"
+        ->merge([$getWireModelType() => $getWireModelValue($locale ?? null)])"
     />
 
-    {{--    @if($isHiveEnabled())--}}
-    {{--        @include('chief-hive::suggest', ['payload' => $getHivePayload($locale ?? null, $this->getId() ? $this : null)])--}}
-    {{--    @endif--}}
+    {{-- @if($isHiveEnabled()) --}}
+    {{-- @include('chief-hive::suggest', ['payload' => $getHivePayload($locale ?? null, $this->getId() ? $this : null)]) --}}
+    {{-- @endif --}}
 </x-chief::form.input.prepend-append>
+
+@include('chief-form::fields._partials.charactercount')